chore: move modules to its own file
