+++ lib/tesla.ex
@@ -1,211 +1,3 @@
-defmodule Tesla.Error do
-  defexception env: nil, stack: [], reason: nil
-
-  def message(%Tesla.Error{env: %{url: url, method: method}, reason: reason}) do
-    "#{inspect(reason)} (#{method |> to_string |> String.upcase()} #{url})"
-  end
-end
-
-defmodule Tesla.Env do
-  @moduledoc """
-  This module defines a `t:Tesla.Env.t/0` struct that stores all data related to request/response.
-
-  ## Fields
-
-  - `:method` - method of request. Example: `:get`
-  - `:url` - request url. Example: `"https://www.google.com"`
-  - `:query` - list of query params.
-    Example: `[{"param", "value"}]` will be translated to `?params=value`.
-    Note: query params passed in url (e.g. `"/get?param=value"`) are not parsed to `query` field.
-  - `:headers` - list of request/response headers.
-    Example: `[{"content-type", "application/json"}]`.
-    Note: request headers are overridden by response headers when adapter is called.
-  - `:body` - request/response body.
-    Note: request body is overridden by response body when adapter is called.
-  - `:status` - response status. Example: `200`
-  - `:opts` - list of options. Example: `[adapter: [recv_timeout: 30_000]]`
-  """
-
-  @type client :: Tesla.Client.t()
-  @type method :: :head | :get | :delete | :trace | :options | :post | :put | :patch
-  @type url :: binary
-  @type param :: binary | [{binary | atom, param}]
-  @type query :: [{binary | atom, param}]
-  @type headers :: [{binary, binary}]
-
-  @type body :: any
-  @type status :: integer | nil
-  @type opts :: keyword
-
-  @type runtime :: {atom, atom, any} | {atom, atom} | {:fn, (t -> t)} | {:fn, (t, stack -> t)}
-  @type stack :: [runtime]
-  @type result :: {:ok, t()} | {:error, any}
-
-  @type t :: %__MODULE__{
-          method: method,
-          query: query,
-          url: url,
-          headers: headers,
-          body: body,
-          status: status,
-          opts: opts,
-          __module__: atom,
-          __client__: client
-        }
-
-  defstruct method: nil,
-            url: "",
-            query: [],
-            headers: [],
-            body: nil,
-            status: nil,
-            opts: [],
-            __module__: nil,
-            __client__: nil
-end
-
-defmodule Tesla.Middleware do
-  @moduledoc """
-  The middleware specification.
-
-  Middleware is an extension of basic `Tesla` functionality. It is a module that must
-  implement `c:Tesla.Middleware.call/3`.
-
-  ## Middleware options
-
-  Options can be passed to middleware in second param of `Tesla.Builder.plug/2` macro:
-
-      plug Tesla.Middleware.BaseUrl, "https://example.com"
-
-  or inside tuple in case of dynamic middleware (`Tesla.client/1`):
-
-      Tesla.client([{Tesla.Middleware.BaseUrl, "https://example.com"}])
-
-  ## Writing custom middleware
-
-  Writing custom middleware is as simple as creating a module implementing `c:Tesla.Middleware.call/3`.
-
-  See `c:Tesla.Middleware.call/3` for details.
-
-  ### Examples
-
-      defmodule MyProject.InspectHeadersMiddleware do
-        @behaviour Tesla.Middleware
-
-        @impl Tesla.Middleware
-        def call(env, next, options) do
-          env
-          |> inspect_headers(options)
-          |> Tesla.run(next)
-          |> inspect_headers(options)
-        end
-
-        defp inspect_headers(env, options) do
-          IO.inspect(env.headers, options)
-        end
-      end
-
-  """
-
-  @doc """
-  Invoked when a request runs.
-
-  - (optionally) read and/or writes request data
-  - calls `Tesla.run/2`
-  - (optionally) read and/or writes response data
-
-  ## Arguments
-
-  - `env` - `Tesla.Env` struct that stores request/response data
-  - `next` - middlewares that should be called after current one
-  - `options` - middleware options provided by user
-  """
-  @callback call(env :: Tesla.Env.t(), next :: Tesla.Env.stack(), options :: any) ::
-              Tesla.Env.result()
-end
-
-defmodule Tesla.Adapter do
-  @moduledoc """
-  The adapter specification.
-
-  Adapter is a module that denormalize request data stored in `Tesla.Env` in order to make
-  request with lower level http client (e.g. `:httpc` or `:hackney`) and normalize response data
-  in order to store it back to `Tesla.Env`. It has to implement `c:Tesla.Adapter.call/2`.
-
-  ## Writing custom adapter
-
-  Create a module implementing `c:Tesla.Adapter.call/2`.
-
-  See `c:Tesla.Adapter.call/2` for details.
-
-  ### Examples
-
-      defmodule MyProject.CustomAdapter do
-        alias Tesla.Multipart
-
-        @behaviour Tesla.Adapter
-
-        @override_defaults [follow_redirect: false]
-
-        @impl Tesla.Adapter
-        def call(env, opts) do
-          opts = Tesla.Adapter.opts(@override_defaults, env, opts)
-
-          with {:ok, {status, headers, body}} <- request(env.method, env.body, env.headers, opts) do
-            {:ok, normalize_response(env, status, headers, body)}
-          end
-        end
-
-        defp request(_method, %Stream{}, _headers, _opts) do
-          {:error, "stream not supported by adapter"}
-        end
-
-        defp request(_method, %Multipart{}, _headers, _opts) do
-          {:error, "multipart not supported by adapter"}
-        end
-
-        defp request(method, body, headers, opts) do
-          :lower_level_http.request(method, body, denormalize_headers(headers), opts)
-        end
-
-        defp denormalize_headers(headers), do: ...
-        defp normalize_response(env, status, headers, body), do: %Tesla.Env{env | ...}
-      end
-
-  """
-
-  @doc """
-  Invoked when a request runs.
-
-  ## Arguments
-
-  - `env` - `Tesla.Env` struct that stores request/response data
-  - `options` - middleware options provided by user
-  """
-  @callback call(env :: Tesla.Env.t(), options :: any) :: Tesla.Env.result()
-
-  @doc """
-  Helper function that merges all adapter options.
-
-  ## Arguments
-
-  - `defaults` (optional) - useful to override lower level http client default configuration
-  - `env` - `Tesla.Env` struct
-  - `opts` - options provided to `Tesla.Builder.adapter/2` macro
-
-  ## Precedence rules
-
-  - config from `opts` overrides config from `defaults` when same key is encountered
-  - config from `env` overrides config from both `defaults` and `opts` when same key is encountered
-  """
-  @spec opts(Keyword.t(), Tesla.Env.t(), Keyword.t()) :: Keyword.t()
-  def opts(defaults \\ [], env, opts) do
-    defaults
-    |> Keyword.merge(opts || [])
-    |> Keyword.merge(env.opts[:adapter] || [])
-  end
-end
-
 defmodule Tesla do
   use Tesla.Builder
