+++ lib/tesla/middleware/follow_redirects.ex
@@ -31,7 +31,7 @@ defmodule Tesla.Middleware.FollowRedirects do
 
   defp redirect(env, next, left) when left == 0 do
     case Tesla.run(env, next) do
-      {:ok, %{status: status} = env} when not (status in @redirect_statuses) ->
+      {:ok, %{status: status} = env} when status not in @redirect_statuses ->
         {:ok, env}
 
       {:ok, _env} ->
